ubuntu-cpc: Initramfs compression should match runtime default (LP: #1744114)
Cloud images created from the 'ubuntu-cpc' project use lzma initramfs
compression for the initial initramfs shipped in the image; this is the
default for livecd-rootfs. When a user updates the initramfs (or installs
a new kernel) on a running instance created from a cloud image they will
get gzip compression, which is the default shipped in initramfs-tools.

With this patch the compression method used to create the initramfs in
the initial cloud images will match the compression menthod used when
the end-user creates/updates an initramfs, namely gzip compression.